Hides tessera field and disables input for now
This field is only for getting the list of public keys right now. And the endpoint is currently on the P2P api at /partyinfo, which won't be open usually and won't have the right CORS settings. In the next version of Tessera, we are adding partyinfo to the 3rd Party api. Once that is done we can re-add this field.

This change doesn't affect sending private transactions. You can still do that by manually adding the keys to the privateFor dropdown and deploying a new contract.